+++ apps/member-profile/app/routes/_public.login.oauth.tsx
@@ -1,11 +1,11 @@
 import { type LoaderFunctionArgs, redirect } from '@remix-run/node';
 import jwt from 'jsonwebtoken';
 
+import { track } from '@oyster/infrastructure/mixpanel';
 import { toTitleCase } from '@oyster/utils';
 
 import { Route } from '@/shared/constants';
 import { ENV } from '@/shared/constants.server';
-import { trackWithoutRequest } from '@/shared/mixpanel.server';
 import { commitSession, getSession, SESSION } from '@/shared/session.server';
 
 // TODO: Add Zod validation here.
@@ -43,8 +43,10 @@ export async function loader({ request }: LoaderFunctionArgs) {
 
   session.set(SESSION.USER_ID, id);
 
-  trackWithoutRequest(id, 'Logged In', {
-    Method: toTitleCase(method) as 'Google' | 'Slack',
+  track({
+    event: 'Logged In',
+    properties: { Method: toTitleCase(method) as 'Google' | 'Slack' },
+    user: id,
   });
 
   const redirectUrl = session.get(SESSION.REDIRECT_URL) || Route['/home'];
