+++ packages/core/src/member-profile.server.ts
@@ -47,4 +47,3 @@ export { updateMember } from './modules/member/use-cases/update-member';
 export { reportException } from './modules/sentry/use-cases/report-exception';
 export { countMessagesSent } from './modules/slack/queries/count-messages-sent';
 export { claimSwagPack } from './modules/swag-pack/use-cases/claim-swag-pack';
-export { getIpAddress } from './shared/utils/ip.utils';
