+++ packages/utils/package.json
@@ -3,15 +3,22 @@
   "version": "0.0.0",
   "private": true,
   "type": "module",
-  "main": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./get-ip-address": "./src/get-ip-address.ts"
+  },
   "scripts": {
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "test": "vitest run",
     "type-check": "tsc --noEmit"
   },
+  "optionalDependencies": {
+    "is-ip": "^5.0.1"
+  },
   "devDependencies": {
     "@oyster/eslint-config": "*",
+    "is-ip": "^5.0.1",
     "nanoid": "^3.0.0",
     "vitest": "^1.3.1"
   }
